refactor(plugin-openclaw): extract daemon host classification from bridge.ts
bridge.ts crossed the 1200 LOC cap for non-grandfathered files with the
same-host dial fallback; address classification is a concern of its own.

Part of #3073
